Update Tutorial5
Added check if the path it is trying to load is null or empty. This removes a major crash.

Added a check if the pathit is trying to load is a file. This removes a runtime error caught by try catch.

Updated code to C# 6.0

Fixed Treeview nodes being cleared when the back button has no parent directory to go to

Added some comments to some of the code